put the more complex string functions into main.py
