+++ src/psssodls/constraints.py
@@ -16,25 +16,10 @@ def vecneq(p, q):
   the two vectors p and q."""
   return 'watchvecneq({}, {})'.format(ellell(p), ellell(q))
 
-def latin_constraints_str(n):
-  """Returns a string containing the latin constraints for a square of size
-  n."""
-  s = '# Latin constraints. \n\n'
-  s += "".join([alldiff([i,'_']) + '\n' + alldiff(['_',i]) + '\n' for i in range(n)])
-  return s + '\n'
-
 def sdk_positions_box(i,j,p):
   row_offset = p * (i-1)
   column_offset = p * (j-1)
   return [(row_offset + r, column_offset + c) for r in range(p) for c in range(p)]
-
-def box_constraints_str(n):
-  p = int(sqrt(n))
-  F = [sdk_positions_box(i,j,p) for i in range(1, p + 1) for j in range(1, p + 1)]
-  s = '# Box constraints. \n\n'
-  for p in F:
-    s += 'alldiff([' + ",".join(ell(q) for q in p) + '])' + '\n'
-  return s + '\n'
 
 def pandiagonal_sum_a(n, w):
   return ",".join([ell([i % n, (i + w) % n]) for i in range(n)])
